Add files via upload
This is a Line Following Robot that works autonomously. We programmed Arduino Mega as the brain of this mobile robot. This robot is made up of an arry of IR sensors, a color sensor, an ultrasonic sensor, L298N motor Driver, Volt and amp display meter, encoder motors, WiFi and SD card Modules. It is a wireless microcontroller based robot. This is our semester project for Micorcontrollers and Embedded System for the Degree of Bachelors of Mechatronics.